Remove the IllegalInstanceVariableInPatternSynonym as a quick hack
The reason being: Instance vars work with the @inst annotation but annotation info is not included in the handling of patterns. A lot of code would have to be changed if @inst were to be properly implemented with patterns.
It is easier to just use the existing code that translates the hiding in Agda.TypeChecking.Rules.Application:173.